feat: 压缩存储空间为32位下标

+++ include/Core/Concepts.hpp
@@ -10,11 +10,11 @@ namespace TDFANN {
 
 template <typename T>
 concept IndexOrList =
-    std::convertible_to<T, size_t> || Graph::IndexList<std::decay_t<T>>;
+    std::convertible_to<T, unsigned> || Graph::IndexList<std::decay_t<T>>;
 
 template <typename Op, typename T>
 concept IndexOrVector =
-    std::convertible_to<Op, size_t> || Vector::DotProductWithVectorType<Op, T>;
+    std::convertible_to<Op, unsigned> || Vector::DotProductWithVectorType<Op, T>;
 
 template <typename Op>
 concept IsTDFG = std::is_same_v<Op, Graph::TDGraphIndexBase> ||
